@@ -1,27 +1,24 @@
-﻿namespace IOGameServer.Application.Models.Components.Collision
+﻿namespace IOGameServer.Application.Models.Components.Collision;
+
+public abstract class CollisionBorder(IGameObject gameObject) : Component(gameObject)
 {
-    public abstract class CollisionBorder : Component
+    protected abstract void HandleReachedBorder();
+
+    public override void Start() { }
+
+    public override void Update(double _)
     {
-        protected CollisionBorder(IGameObject gameObject) : base(gameObject) { }
-
-        protected abstract void HandleReachedBorder();
-
-        public override void Start() { }
-
-        public override void Update(double _)
+        if (HasReachedBorder())
         {
-            if (HasReachedBorder())
-            {
-                HandleReachedBorder();
-            }
+            HandleReachedBorder();
         }
+    }
 
-        private bool HasReachedBorder()
-        {
-            return GameObject.X <= 0
-                || GameObject.Y <= 0
-                || GameObject.X >= GameObject.Game.Settings.MapSize
-                || GameObject.Y >= GameObject.Game.Settings.MapSize;
-        }
+    private bool HasReachedBorder()
+    {
+        return GameObject.X <= 0
+            || GameObject.Y <= 0
+            || GameObject.X >= GameObject.Game.Settings.MapSize
+            || GameObject.Y >= GameObject.Game.Settings.MapSize;
     }
 }
