@@ -1,13 +1,10 @@
-﻿namespace IOGameServer.Application.Models.Components.Collision
-{
-    public sealed class CollisionBorderStayInside : CollisionBorder
-    {
-        public CollisionBorderStayInside(IGameObject gameObject) : base(gameObject) { }
+﻿namespace IOGameServer.Application.Models.Components.Collision;
 
-        protected override void HandleReachedBorder()
-        {
-            GameObject.X = Math.Max(5, Math.Min(GameObject.Game.Settings.MapSize, GameObject.X));
-            GameObject.Y = Math.Max(5, Math.Min(GameObject.Game.Settings.MapSize, GameObject.Y));
-        }
+public sealed class CollisionBorderStayInside(IGameObject gameObject) : CollisionBorder(gameObject)
+{
+    protected override void HandleReachedBorder()
+    {
+        GameObject.X = Math.Max(5, Math.Min(GameObject.Game.Settings.MapSize, GameObject.X));
+        GameObject.Y = Math.Max(5, Math.Min(GameObject.Game.Settings.MapSize, GameObject.Y));
     }
 }
