@@ -1,27 +1,24 @@
-﻿namespace IOGameServer.Application.Models.Components.Collision
+﻿namespace IOGameServer.Application.Models.Components.Collision;
+
+public sealed class CollisionObject(IGameObject gameObject) : Component(gameObject)
 {
-    public sealed class CollisionObject : Component
+    public required int Radius { get; set; }
+    Damageable.Damageable Damageable { get; set; }
+
+    public override void Start()
     {
-        public required int Radius { get; set; }
-        Damageable.Damageable Damageable { get; set; }
+        Damageable = GameObject.GetComponent<Damageable.Damageable>();
+    }
 
-        public CollisionObject(IGameObject gameObject) : base(gameObject) { }
+    public override void Update(double distance) { }
 
-        public override void Start()
-        {
-            Damageable = GameObject.GetComponent<Damageable.Damageable>();
-        }
+    public bool Collided(IGameObject gameObject)
+    {
+        return GameObject.DistanceTo(gameObject) <= Radius + gameObject.GetComponent<CollisionObject>().Radius;
+    }
 
-        public override void Update(double distance) { }
-
-        public bool Collided(IGameObject gameObject)
-        {
-            return GameObject.DistanceTo(gameObject) <= Radius + gameObject.GetComponent<CollisionObject>().Radius;
-        }
-
-        public void HandleCollision(IGameObject gameObject)
-        {
-            Damageable?.HandleDamage(gameObject);
-        }
+    public void HandleCollision(IGameObject gameObject)
+    {
+        Damageable?.HandleDamage(gameObject);
     }
 }
