@@ -1,30 +1,27 @@
 ﻿using IOGameServer.Application.Models.GameObjects;
 
-namespace IOGameServer.Application.Models.Components.Health
+namespace IOGameServer.Application.Models.Components.Health;
+
+public sealed class Health(Player gameObject) : Component(gameObject)
 {
-    public sealed class Health : Component
+    public required int HP { get; set; }
+
+    public override void Start() { }
+
+    public override void Update(double _) { }
+
+    public void Heal(int amount)
     {
-        public required int HP { get; set; }
+        HP += amount;
+    }
 
-        public Health(Player gameObject) : base(gameObject) { }
+    public void TakeDamage(int amount)
+    {
+        HP -= amount;
 
-        public override void Start() { }
-
-        public override void Update(double _) { }
-
-        public void Heal(int amount)
+        if (HP <= 0)
         {
-            HP += amount;
-        }
-
-        public void TakeDamage(int amount)
-        {
-            HP -= amount;
-
-            if (HP <= 0)
-            {
-                GameObject.RemoveMe();
-            }
+            GameObject.RemoveMe();
         }
     }
 }
