@@ -1,10 +1,9 @@
-﻿namespace IOGameServer.Application.Models.Components
-{
-    public interface IComponent
-    {
-        IGameObject GameObject { get; }
+﻿namespace IOGameServer.Application.Models.Components;
 
-        void Start();
-        void Update(double distance);
-    }
+public interface IComponent
+{
+    IGameObject GameObject { get; }
+
+    void Start();
+    void Update(double distance);
 }
