@@ -5,30 +5,29 @@ using IOGameServer.Application.Models.Components.Movement;
 using IOGameServer.Application.Models.Components.Spawn;
 using IOGameServer.Application.Models.Inputs;
 
-namespace IOGameServer.Application.Models.GameObjects
+namespace IOGameServer.Application.Models.GameObjects;
+
+public sealed class Bullet : GameObject
 {
-    public sealed class Bullet : GameObject
+    public Bullet(Game game, IGameObject shotByPlayer, int x, int y, double direction) : base(game)
     {
-        public Bullet(Game game, IGameObject shotByPlayer, int x, int y, double direction) : base(game)
+        AddComponent(new SpawnFixed(this, x, y));
+        AddComponent(new CollisionBorderRemoveOnTouch(this));
+        AddComponent(new CollisionObject(this) { Radius = Game.Settings.BulletRadius });
+        AddComponent(new Damageable(this) { ShotByPlayer = shotByPlayer, });
+        AddComponent(new MovementNormal(this)
         {
-            AddComponent(new SpawnFixed(this, x, y));
-            AddComponent(new CollisionBorderRemoveOnTouch(this));
-            AddComponent(new CollisionObject(this) { Radius = Game.Settings.BulletRadius });
-            AddComponent(new Damageable(this) { ShotByPlayer = shotByPlayer, });
-            AddComponent(new MovementNormal(this)
-            {
-                Speed = Game.Settings.BulletSpeed,
-                Direction = direction
-            });
-        }
-
-        public override void HandleInput(IInput input) { } // bullets take no input
-
-        public UpdateModel.Bullet GetClientModel() => new()
-        {
-            Id = Id,
-            X = X,
-            Y = Y
-        };
+            Speed = Game.Settings.BulletSpeed,
+            Direction = direction
+        });
     }
+
+    public override void HandleInput(IInput input) { } // bullets take no input
+
+    public UpdateModel.Bullet GetClientModel() => new()
+    {
+        Id = Id,
+        X = X,
+        Y = Y
+    };
 }
