@@ -1,4 +1,3 @@
-﻿namespace IOGameServer.Application.Models.Inputs
-{
-    public interface IInput { }
-}
+﻿namespace IOGameServer.Application.Models.Inputs;
+
+public interface IInput { }
