@@ -1,17 +1,16 @@
-﻿namespace IOGameServer.Application.Settings
+﻿namespace IOGameServer.Application.Settings;
+
+public sealed class GameSettings
 {
-    public sealed class GameSettings
-    {
-        public int TotalPlayersPerGame { get; init; }
-        public int MapSize { get; init; }
-        public int PlayerRadius { get; init; }
-        public int PlayerMaxHP { get; init; }
-        public int PlayerSpeed { get; init; }
-        public double PlayerFireCooldown { get; init; }
-        public int BulletRadius { get; init; }
-        public int BulletSpeed { get; init; }
-        public int BulletDamage { get; init; }
-        public int ScoreBulletHit { get; init; }
-        public int ScorePerSecond { get; init; }
-    }
+    public int TotalPlayersPerGame { get; init; }
+    public int MapSize { get; init; }
+    public int PlayerRadius { get; init; }
+    public int PlayerMaxHP { get; init; }
+    public int PlayerSpeed { get; init; }
+    public double PlayerFireCooldown { get; init; }
+    public int BulletRadius { get; init; }
+    public int BulletSpeed { get; init; }
+    public int BulletDamage { get; init; }
+    public int ScoreBulletHit { get; init; }
+    public int ScorePerSecond { get; init; }
 }
