@@ -3,59 +3,53 @@ using IOGameServer.Helpers;
 using Microsoft.AspNetCore.SignalR;
 using IOGameServer.Application.Models.Inputs.Player;
 
-namespace IOGameServer.Hubs
+namespace IOGameServer.Hubs;
+
+public sealed class GameHub(GameService gameService) : Hub<IGameHub>
 {
-    public sealed class GameHub : Hub<IGameHub>
+    private const string GameIdKey = "groupId";
+    private const string PlayerIdKey = "playerId";
+
+    private readonly GameService _gameService = gameService;
+
+    public async Task JoinGame(string username)
     {
-        private const string GameIdKey = "groupId";
-        private const string PlayerIdKey = "playerId";
+        username = await UsernameValidator.Validate(username);
 
-        private readonly GameService _gameService;
+        var (game, player) = _gameService
+            .AddPlayer(username, Context.ConnectionId);
 
-        public GameHub(GameService gameService)
-        {
-            _gameService = gameService;
-        }
+        Context.Items[GameIdKey] = game.Id;
+        Context.Items[PlayerIdKey] = player.Id;
+    }
 
-        public async Task JoinGame(string username)
-        {
-            username = await UsernameValidator.Validate(username);
+    public void ChangeDirection(double direction)
+    {
+        _gameService
+            .GetGame(GetGameId())?
+            .GetPlayer(GetPlayerId())?
+            .HandleInput(new DirectionInput
+            {
+                Direction = direction
+            });
+    }
 
-            var (game, player) = _gameService
-                .AddPlayer(username, Context.ConnectionId);
+    public override async Task OnDisconnectedAsync(Exception exception)
+    {
+        _gameService
+            .RemovePlayer(GetGameId(), GetPlayerId());
 
-            Context.Items[GameIdKey] = game.Id;
-            Context.Items[PlayerIdKey] = player.Id;
-        }
+        await base.OnDisconnectedAsync(exception);
+    }
+    
+    private string GetPlayerId()
+    {
+        return (string)Context.Items[PlayerIdKey];
+    }
 
-        public void ChangeDirection(double direction)
-        {
-            _gameService
-                .GetGame(GetGameId())?
-                .GetPlayer(GetPlayerId())?
-                .HandleInput(new DirectionInput
-                {
-                    Direction = direction
-                });
-        }
+    private string GetGameId()
+    {
+        return (string)Context.Items[GameIdKey];
 
-        public override async Task OnDisconnectedAsync(Exception exception)
-        {
-            _gameService
-                .RemovePlayer(GetGameId(), GetPlayerId());
-
-            await base.OnDisconnectedAsync(exception);
-        }
-        
-        private string GetPlayerId()
-        {
-            return (string)Context.Items[PlayerIdKey];
-        }
-
-        private string GetGameId()
-        {
-            return (string)Context.Items[GameIdKey];
-
-        }
     }
 }
