@@ -1,10 +1,9 @@
 ﻿using IOGameServer.Hubs.ClientModels;
 
-namespace IOGameServer.Hubs
+namespace IOGameServer.Hubs;
+
+public interface IGameHub
 {
-    public interface IGameHub
-    {
-        Task GameOver();
-        Task GameUpdate(UpdateModel update);
-    }
+    Task GameOver();
+    Task GameUpdate(UpdateModel update);
 }
